requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.5.14
distro==1.8.0
dnspython==2.4.2
exceptiongroup==1.1.3
Flask==2.3.2
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langchain==0.0.189
langchainplus-sdk==0.0.20
langsmith==0.0.30
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
numexpr==2.8.7
numpy==1.26.1
openai==0.28.1
openapi-schema-pydantic==1.2.4
packaging==23.2
pinecone-client==2.2.4
pydantic==1.10.13
pymongo==4.5.0
pypdf==3.17.0
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
tenacity==8.2.3
tiktoken==0.5.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==1.26.18
Werkzeug==3.0.1
yarl==1.9.2
